requirements.txt
dataclasses==0.6
Django==2.2.7
djangorestframework==3.10.3
djongo==1.2.36
joblib==0.14.0
numpy==1.17.4
pandas==0.25.3
psycopg2==2.8.4
Pygments==2.5.1
pymongo==3.9.0
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
sqlparse==0.2.4
